Updating the Pivot to remove the hardcoded content rendering.
